+++ src/main/java/com/bridgeX/user/dto/SignupRequest.java
@@ -37,8 +37,8 @@ public class SignupRequest {
 	private String email;
 
 	@Min(value = 1, message = "[SERVER] 키는 0 이하일 수 없습니다.")
-	private Integer height;
+	private Double height;
 
 	@Min(value = 1, message = "[SERVER] 체중은 0 이하일 수 없습니다.")
-	private Integer weight;
+	private Double weight;
 }
