+++ src/main/java/com/bridgeX/user/dto/UserBodyInfoRequest.java
@@ -12,21 +12,21 @@ import lombok.Setter;
 @Setter
 public class UserBodyInfoRequest {
 	@Min(1)
-	private Integer height;
+	private Double height;
 
 	@Min(1)
-	private Integer weight;
+	private Double weight;
 
 	@NotNull
 	private UserGender gender; // e-num { MALE, FEMALE }
 
-	private Integer user_grip;
+	private Double user_grip;
 	
-	private Integer user_flex;
+	private Double user_flex;
 	
 	private Integer user_situp;
 	
-	private Integer user_jump;
+	private Double user_jump;
 	
 	// TODO: Add Other fields
 }
